New model

+++ moving_cells_utils/uopvp_solver.py
@@ -354,10 +354,10 @@ class UOPVPSolver:
         available_time = self.t_max - sum(t_route)
         service_intervals = []
         for m in range(len(t_route)):
-            s = (available_time - sum(service_intervals[:m])) / (len(route) - m)
-            left, right, t_lr = route[m], route[m+1], t_route[m]
+            s = int((available_time - sum(service_intervals[:m])) / (len(route) - m))
+            left, right, t_lr = route[m], route[m+1], int(t_route[m])
 
-            time_to_m = sum(service_intervals[:m]) + sum(t_route[:m])
+            time_to_m = int(sum(service_intervals[:m]) + sum(t_route[:m]))
             start, stop = time_to_m, time_to_m+t_lr
             profit_avg_i = sum(self.find_profit_and_beta(route[m][0], route[m][1], t)[0]
                                for t in range(start+s, stop+s))/t_lr
@@ -372,8 +372,8 @@ class UOPVPSolver:
             while (not ((profit_avg_i - profit_avg_j)**2 < self.optimizer_minimum and profit_chg_i - profit_chg_j < 0)
                    and step < self.optimizer_steps and s > 0):
                 print(step)
-                s = max(0, s + int(self.optimizer_weight * (profit_avg_i - profit_avg_j) +
-                                   self.optimizer_weight**2 * (profit_chg_i**2 - profit_chg_j**2)))
+                s = max(0, int(s + int(self.optimizer_weight * (profit_avg_i - profit_avg_j) +
+                                       self.optimizer_weight**2 * (profit_chg_i**2 - profit_chg_j**2))))
                 profit_avg_i = sum(self.find_profit_and_beta(route[m][0], route[m][1], t)[0]
                                    for t in range(start + s, stop + s)) / t_lr
                 profit_avg_j = sum(self.find_profit_and_beta(route[m + 1][0], route[m + 1][1], t)[0]
